requirements.txt
requests
python-telegram-bot
pandas
sentence-transformers
scikit-learn
python-dotenv
openpyxl
pyTelegramBotAPI